requirements.txt
bottle<0.13.0,>=0.12.13
gunicorn<20.0.0,>=19.7.1
mecab-python3<0.8.0,>=0.7.0
